+++ gerar-dados.h
@@ -2,7 +2,7 @@
   #define __GERAR_DADOS_H__
 
   extern const char *nome_base[];
-  extern const float preco_base[];
+  extern const double preco_base[];
 
   int get_size_nome(); 
   int get_size_preco();
